Fixed some compiler warnings in SQLoxx tests.
git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/sqloxx@2184 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98

+++ tests/table_iterator_tests.cpp
@@ -74,7 +74,7 @@ TEST_FIXTURE
 
 	size_t count = 0;
 	for (DerivedPOHandleIter it(*pdbc); it != null_iter; ++it) ++count;
-	CHECK_EQUAL(count, 5);
+	CHECK_EQUAL(count, static_cast<size_t>(5));
 
 	int max = 0;
 	for (DerivedPOHandleIter it(*pdbc); it != null_iter; ++it)
@@ -101,7 +101,7 @@ TEST_FIXTURE
 	);
 	size_t count = 0;
 	for ( ; it != null_iter; ++it) ++count;
-	CHECK_EQUAL(count, 3);
+	CHECK_EQUAL(count, static_cast<size_t>(3));
 }
 
 TEST_FIXTURE
